Make Expr and Stmt implementations regular classes
Data classes' equality breaks the interpreter. Two different tokens
(variables) in the same line would be interpreted as the same. Even
the variable token repeated in "for" syntactic sugar expansion was
considered the same.

This messes with the scope resolution, where a token from different
scope would override a previous one.